twistlock-prefect/pyproject.toml
[project]
name = "twistlock-prefect"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "prefect>=2.14",
    "boto3",
    "psycopg2-binary",
    "python-dotenv",
    "requests",
]

[tool.uv]
dev-dependencies = [
    "pytest",
    "ruff",
]

[tool.ruff]
line-length = 100
target-version = "py313"
